requirements.txt
streamlit==1.17.0
streamlit-folium==0.10.0
plotly==5.12.0
numpy==1.24.1
folium==0.14.0
matplotlib==3.6.3
haversine==2.7.0
streamlit-folium==0.10.0
Pillow==9.4.0